Refactor folder structure of Case API routes
Instead of all routes in one big folder, now moved in subfolders that sort of reflect the API hierarchy as well

Also created a DiscretionaryAPIFormat class holding some of the CaseModelResponse, and added StartCaseReponse to the StartCaseFormat.
