+++ src/main/scala/org/cafienne/service/akkahttp/cases/CasesRoute.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.cafienne.service.akkahttp.cases.route
+package org.cafienne.service.akkahttp.cases
 
 import akka.http.scaladsl.model._
 import akka.http.scaladsl.server.Route
