+++ src/main/scala/org/cafienne/service/akkahttp/cases/deprecated/DeprecatedCaseTeamRoute.scala
@@ -15,13 +15,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.cafienne.service.akkahttp.cases.route.deprecated
+package org.cafienne.service.akkahttp.cases.deprecated
 
 import akka.http.scaladsl.server.Route
 import org.cafienne.cmmn.actorapi.command.team.DeprecatedUpsert
 import org.cafienne.cmmn.actorapi.command.team.removemember.{RemoveCaseTeamTenantRole, RemoveCaseTeamUser}
-import org.cafienne.service.akkahttp.cases.model.CaseTeamAPI.Compatible._
-import org.cafienne.service.akkahttp.cases.route.CasesRoute
+import org.cafienne.service.akkahttp.cases.CasesRoute
+import org.cafienne.service.akkahttp.cases.team.CaseTeamAPI.Compatible._
 import org.cafienne.system.CaseSystem
 
 class DeprecatedCaseTeamRoute(override val caseSystem: CaseSystem) extends CasesRoute {
