+++ src/main/scala/org/cafienne/service/akkahttp/cases/documentation/CaseDocumentationRoute.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.cafienne.service.akkahttp.cases.route
+package org.cafienne.service.akkahttp.cases.documentation
 
 import akka.http.scaladsl.server.Route
 import io.swagger.v3.oas.annotations.enums.ParameterIn
@@ -24,6 +24,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.security.SecurityRequirement
 import io.swagger.v3.oas.annotations.{Operation, Parameter}
 import org.cafienne.service.akkahttp.Headers
+import org.cafienne.service.akkahttp.cases.CasesRoute
 import org.cafienne.system.CaseSystem
 
 import javax.ws.rs._
