+++ src/main/scala/org/cafienne/service/akkahttp/cases/plan/PlanItemRoute.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.cafienne.service.akkahttp.cases.route
+package org.cafienne.service.akkahttp.cases.plan
 
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.Route
@@ -24,22 +24,20 @@ import io.swagger.v3.oas.annotations.media.Schema
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.security.SecurityRequirement
 import io.swagger.v3.oas.annotations.{Operation, Parameter}
-import org.cafienne.json.Value
+import org.cafienne.cmmn.actorapi.command.plan.MakePlanItemTransition
+import org.cafienne.cmmn.instance.Transition
 import org.cafienne.service.akkahttp.Headers
-import org.cafienne.service.akkahttp.debug.ModelEventsReader
+import org.cafienne.service.akkahttp.cases.CasesRoute
 import org.cafienne.system.CaseSystem
 
 import javax.ws.rs._
-import scala.util.{Failure, Success}
 
 @SecurityRequirement(name = "oauth2", scopes = Array("openid"))
 @Path("/cases")
-class CaseHistoryRoute(override val caseSystem: CaseSystem) extends CaseEventsBaseRoute {
-  override def routes: Route = concat(getCaseEvents, getPlanHistory, getPlanItemHistory)
+class PlanItemRoute(override val caseSystem: CaseSystem) extends CasesRoute {
+  override def routes: Route = concat(getPlanItems, getPlanItem, makePlanItemTransition)
 
-  val modelEventsReader = new ModelEventsReader(caseSystem)
-
-  @Path("/{caseInstanceId}/history/events")
+  @Path("/{caseInstanceId}/planitems")
   @GET
   @Operation(
     summary = "Get the plan items of a case",
@@ -55,66 +53,66 @@ class CaseHistoryRoute(override val caseSystem: CaseSystem) extends CaseEventsBa
     )
   )
   @Produces(Array("application/json"))
-  def getCaseEvents: Route = get {
-    caseEventsSubRoute { caseEvents =>
-      path("history" / "events") {
-        onComplete(caseEvents.eventList()) {
-          case Success(events) => completeJson(Value.convert(events.map(_.event.rawJson())))
-          case Failure(err) => complete(StatusCodes.NotFound, err)
-        }
+  def getPlanItems: Route = get {
+    caseInstanceSubRoute { (user, caseInstanceId) =>
+      path("planitems") {
+        runListQuery(caseQueries.getPlanItems(caseInstanceId, user))
       }
     }
   }
 
-  @Path("/{caseInstanceId}/history/planitems")
+  @Path("/{caseInstanceId}/planitems/{planItemId}")
   @GET
   @Operation(
-    summary = "Get the plan items of a case",
-    description = "Get the plan items of the specified case instance",
+    summary = "Get a plan item of a case",
+    description = "Get a plan item of the specified case instance",
     tags = Array("case plan"),
     parameters = Array(
       new Parameter(name = "caseInstanceId", description = "Unique id of the case instance", in = ParameterIn.PATH, schema = new Schema(implementation = classOf[String]), required = true),
+      new Parameter(name = "planItemId", description = "Unique id of the plan item", in = ParameterIn.PATH, schema = new Schema(implementation = classOf[String]), required = true),
       new Parameter(name = Headers.CASE_LAST_MODIFIED, description = "Get after events have been processed", in = ParameterIn.HEADER, schema = new Schema(implementation = classOf[String]), required = false),
-    ),
-    responses = Array(
-      new ApiResponse(description = "Plan items found", responseCode = "200"),
-      new ApiResponse(description = "No plan items found based on the query params", responseCode = "404")
-    )
-  )
-  @Produces(Array("application/json"))
-  def getPlanHistory: Route = get {
-    caseEventsSubRoute { caseEvents =>
-      path("history" / "planitems") {
-        onComplete(caseEvents.casePlanHistory()) {
-          case Success(items) => completeJson(Value.convert(items.map(_.toValue)))
-          case Failure(t) => handleFailure(t)
-        }
-      }
-    }
-  }
-
-  @Path("/{caseInstanceId}/history/planitems/{planItemId}")
-  @GET
-  @Operation(
-    summary = "Get history of a plan item in a case by planItemId",
-    description = "Get history of a plan item in the specified case instance",
-    tags = Array("case plan"),
-    parameters = Array(
-      new Parameter(name = "caseInstanceId", description = "Unique id of the case instance", in = ParameterIn.PATH, schema = new Schema(implementation = classOf[String]), required = true),
-      new Parameter(name = "planItemId", description = "Unique id of the plan item (cannot be the plan item name, must be the id)", in = ParameterIn.PATH, schema = new Schema(implementation = classOf[String]), required = true),
     ),
     responses = Array(
       new ApiResponse(description = "Plan item found", responseCode = "200"),
-      new ApiResponse(description = "No plan item found", responseCode = "404")
+      new ApiResponse(description = "Plan item not found", responseCode = "404")
     )
   )
   @Produces(Array("application/json"))
-  def getPlanItemHistory: Route = get {
-    caseEventsSubRoute { caseEvents =>
-      path("history" / "planitems" / Segment) { planItemId =>
-        onComplete(caseEvents.planitemHistory(planItemId)) {
-          case Success(value) => completeJson(value)
-          case Failure(t) => handleFailure(t)
+  def getPlanItem: Route = get {
+    caseInstanceSubRoute { (user, caseInstanceId) =>
+      path("planitems" / Segment) {
+        planItemId => runQuery(caseQueries.getPlanItem(planItemId, user))
+      }
+    }
+  }
+
+  @Path("/{caseInstanceId}/planitems/{identifier}/{transition}")
+  @POST
+  @Operation(
+    summary = "Apply a transition on a plan item",
+    description = "Applies a transition to all plan items in the case that have the identifier. If it is the planItemId there will be only one instance. If planItemName is given, more instances may be found",
+    tags = Array("case plan"),
+    parameters = Array(
+      new Parameter(name = "caseInstanceId", description = "Unique id of the case instance", in = ParameterIn.PATH, schema = new Schema(implementation = classOf[String]), required = true),
+      new Parameter(name = "identifier", description = "Identifier for the plan item; either a plan item id or a plan item name", in = ParameterIn.PATH, schema = new Schema(implementation = classOf[String]), required = true),
+      new Parameter(name = "transition", description = "Transition to apply", in = ParameterIn.PATH,
+        schema = new Schema(implementation = classOf[String], allowableValues = Array("complete", "close", "create", "enable", "disable", "exit", "fault", "manualStart", "occur", "parentResume", "parentSuspend", "reactivate", "reenable", "resume", "start", "suspend", "terminate")),
+        required = true),
+    ),
+    responses = Array(
+      new ApiResponse(description = "Transition applied successfully", responseCode = "202"),
+      new ApiResponse(description = "Case not found", responseCode = "404"),
+    )
+  )
+  @Produces(Array("application/json"))
+  def makePlanItemTransition: Route = post {
+    caseInstanceSubRoute { (user, caseInstanceId) =>
+      path("planitems" / Segment / Segment) { (planItemId, transitionString) =>
+        val transition = Transition.getEnum(transitionString)
+        if (transition == null) {
+          complete(StatusCodes.BadRequest, "Transition " + transition + " is not valid")
+        } else {
+          askCase(user, caseInstanceId, caseMember => new MakePlanItemTransition(caseMember, caseInstanceId, planItemId, transition))
         }
       }
     }
